# public/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GPT Agent Platform API",
    "version": "1.0.0",
    "description": "API для платформы обучения ИИ-агентов"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Основной сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/agents": {
      "get": {
        "summary": "Получение списка ИИ-агентов",
        "description": "Возвращает список ИИ-агентов организации с поддержкой фильтрации и пагинации",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Поиск по имени или описанию агента"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "draft"
              ]
            },
            "description": "Фильтр по статусу агента"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Номер страницы (начиная с 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Количество агентов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ со списком агентов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "agents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "description": "Имя агента"
                          },
                          "description": {
                            "type": "string",
                            "description": "Описание агента"
                          },
                          "instructions": {
                            "type": "string",
                            "description": "Основные инструкции для агента"
                          },
                          "model": {
                            "type": "string",
                            "description": "Модель ИИ (gpt-4, gpt-3.5-turbo и т.д.)"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "inactive",
                              "draft"
                            ]
                          },
                          "welcomeMessage": {
                            "type": "string",
                            "description": "Приветственное сообщение"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Общее количество агентов"
                    },
                    "page": {
                      "type": "integer",
                      "description": "Текущая страница"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Количество агентов на странице"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      },
      "post": {
        "summary": "Создание нового ИИ-агента",
        "description": "Создает нового ИИ-агента с указанными параметрами",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "instructions"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "Имя агента"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Описание агента"
                  },
                  "instructions": {
                    "type": "string",
                    "minLength": 10,
                    "description": "Основные инструкции для агента"
                  },
                  "model": {
                    "type": "string",
                    "default": "gpt-4",
                    "description": "Модель ИИ"
                  },
                  "welcomeMessage": {
                    "type": "string",
                    "description": "Приветственное сообщение"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive",
                      "draft"
                    ],
                    "default": "draft",
                    "description": "Статус агента"
                  }
                },
                "example": {
                  "name": "Менеджер по продажам",
                  "description": "Агент для обработки запросов клиентов",
                  "instructions": "Вы - профессиональный менеджер по продажам. Будьте вежливы, информативны и помогайте клиентам.",
                  "model": "gpt-4",
                  "welcomeMessage": "Здравствуйте! Чем я могу вам помочь?",
                  "status": "active"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Агент успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "agent": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "instructions": {
                          "type": "string"
                        },
                        "model": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "welcomeMessage": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неверные входные данные"
          },
          "401": {
            "description": "Не авторизован"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
...
  },
...
}